use size_t instead of int

+++ eltonfs/xdr/bin_encoding.h
@@ -1,6 +1,7 @@
+#include <linux/types.h>
 
-int bin_encoder_init(struct xdr_encoder *enc, char *buff, int len);
-int bin_decoder_init(struct xdr_decoder *dec, char *buff, int len);
+int bin_encoder_init(struct xdr_encoder *enc, char *buff, size_t len);
+int bin_decoder_init(struct xdr_decoder *dec, char *buff, size_t len);
 
 #ifdef ELTONFS_UNIT_TEST
 void test_xdr_bin(void);
